Add a test for deleting a quote
